requirements.txt
Scrapy~=2.5.0
itemloaders~=1.0.4
itemadapter~=0.2.0
sshtunnel~=0.4.0
pymongo~=3.11.3
python-dotenv~=0.17.0